@@ -21,47 +21,51 @@
     "update-rust-bindings": "cd src-tauri && cargo test"
   },
   "devDependencies": {
-    "@fullhuman/postcss-purgecss": "^6.0.0",
-    "@iconify-json/ic": "^1.2.1",
-    "@iconify-json/mdi": "^1.2.1",
-    "@octokit/plugin-retry": "^7.1.2",
-    "@octokit/plugin-throttling": "^9.3.2",
-    "@octokit/rest": "^21.0.2",
+    "@iconify-json/ic": "^1.2.2",
+    "@iconify-json/mdi": "^1.2.3",
+    "@octokit/plugin-retry": "^7.1.4",
+    "@octokit/plugin-throttling": "^9.4.0",
+    "@octokit/rest": "^21.1.1",
     "@popperjs/core": "^2.11.7",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
-    "@tauri-apps/cli": "^1.6.2",
-    "@testing-library/svelte": "^5.2.3",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tauri-apps/cli": "^2.4.0",
+    "@testing-library/svelte": "^5.2.7",
     "@tsconfig/svelte": "^5.0.4",
-    "@vitest/coverage-v8": "^2.1.3",
+    "@vitest/coverage-v8": "^3.0.8",
     "ansi-to-span": "^0.0.1",
-    "autoprefixer": "^10.4.20",
-    "classnames": "^2.5.1",
+    "autoprefixer": "^10.4.21",
     "escape-html": "^1.0.3",
-    "execa": "^9.4.0",
-    "flowbite": "^2.5.2",
-    "flowbite-svelte": "0.47.0",
-    "flowbite-svelte-icons": "1.6.2",
-    "jsdom": "^25.0.1",
-    "postcss": "^8.4.47",
+    "flowbite": "^3.1.2",
+    "flowbite-svelte": "0.48.4",
+    "flowbite-svelte-icons": "2.0.2",
+    "jsdom": "^26.0.0",
+    "postcss": "^8.5.3",
     "postcss-load-config": "^6.0.1",
-    "prettier": "^3.3.3",
-    "prettier-plugin-svelte": "^3.2.7",
-    "svelte": "^4.2.19",
-    "svelte-i18n": "^4.0.0",
+    "prettier": "^3.5.3",
+    "prettier-plugin-svelte": "^3.3.3",
+    "svelte": "^5.22.6",
+    "svelte-i18n": "^4.0.1",
+    "svelte-navigator": "^3.2.2",
     "svelte-preprocess": "^6.0.3",
     "tailwindcss": "^3.4.13",
-    "typescript": "^5.6.3",
-    "unplugin-icons": "^0.19.3",
-    "vite": "^5.4.9",
-    "vitest": "^2.1.3"
+    "typescript": "^5.8.2",
+    "unplugin-icons": "^22.1.0",
+    "vite": "^6.2.1",
+    "vitest": "^3.0.8"
   },
   "dependencies": {
-    "@tauri-apps/api": "^1.6.0",
-    "@types/semver": "^7.5.8",
-    "country-flag-emoji-polyfill": "^0.1.8",
-    "http-server": "^14.1.1",
-    "semver": "^7.6.3",
-    "svelte-navigator": "^3.2.2"
+    "@tauri-apps/api": "^2.2.0",
+    "@tauri-apps/plugin-clipboard-manager": "^2.0.0",
+    "@tauri-apps/plugin-dialog": "~2",
+    "@tauri-apps/plugin-fs": "~2",
+    "@tauri-apps/plugin-global-shortcut": "~2",
+    "@tauri-apps/plugin-http": "~2",
+    "@tauri-apps/plugin-notification": "~2",
+    "@tauri-apps/plugin-os": "~2",
+    "@tauri-apps/plugin-process": "~2",
+    "@tauri-apps/plugin-shell": "~2",
+    "@tauri-apps/plugin-updater": "~2",
+    "country-flag-emoji-polyfill": "^0.1.8"
   },
   "packageManager": "yarn@1.22.19"
 }
