@@ -4,7 +4,7 @@ use std::{
 };
 
 use log::info;
-use tauri::{api::path::config_dir, Manager};
+use tauri::{Emitter, Manager};
 use walkdir::WalkDir;
 
 use crate::{
@@ -74,19 +74,22 @@ pub async fn uninstall_game(
     .map_err(|_| {
       CommandError::GameManagement("Unable to persist game installation status".to_owned())
     })?;
-  app_handle.emit_all("gameUninstalled", {})?;
+  app_handle.emit("gameUninstalled", {})?;
   Ok(())
 }
 
 #[tauri::command]
-pub async fn reset_game_settings(game_name: String) -> Result<(), CommandError> {
-  let config_dir = match config_dir() {
-    None => {
+pub async fn reset_game_settings(
+  app_handle: tauri::AppHandle,
+  game_name: String,
+) -> Result<(), CommandError> {
+  let config_dir = match app_handle.path().config_dir() {
+    Ok(path) => path,
+    Err(_) => {
       return Err(CommandError::GameManagement(
         "Could not determine game config directory".to_owned(),
       ))
     }
-    Some(path) => path,
   };
 
   let path_to_settings = config_dir
@@ -108,7 +111,7 @@ pub async fn reset_game_settings(game_name: String) -> Result<(), CommandError> 
 
 fn get_saves_highest_milestone(
   path: &PathBuf,
-  milestones: &Vec<MilestoneCriteria>,
+  milestones: &[MilestoneCriteria],
 ) -> Option<(String, i32)> {
   // Read the file's bytes and generate a list of all completed tasks
   let mut tasks: HashMap<u8, GameTaskStatus> = HashMap::new();
@@ -149,18 +152,18 @@ fn get_saves_highest_milestone(
   // Iterate through the milestones backwards
   for (index, milestone) in milestones.iter().rev().enumerate() {
     for task_id in &milestone.introduced {
-      if tasks.contains_key(&task_id) && tasks[&task_id].introduced {
+      if tasks.contains_key(task_id) && tasks[task_id].introduced {
         return Some((milestone.name.to_owned(), (milestones.len() - index) as i32));
       }
     }
     for task_id in &milestone.completed {
-      if tasks.contains_key(&task_id) && tasks[&task_id].completed {
+      if tasks.contains_key(task_id) && tasks[task_id].completed {
         return Some((milestone.name.to_owned(), (milestones.len() - index) as i32));
       }
     }
   }
 
-  return None;
+  None
 }
 
 // Returns the most significant milestone in the game the user has achieved
@@ -169,7 +172,10 @@ fn get_saves_highest_milestone(
 //
 // Otherwise, it will default to a default picture (geyser)
 #[tauri::command]
-pub async fn get_furthest_game_milestone(game_name: String) -> Result<String, CommandError> {
+pub async fn get_furthest_game_milestone(
+  app_handle: tauri::AppHandle,
+  game_name: String,
+) -> Result<String, CommandError> {
   // TODO - currently only checking Jak 1
   // TODO - It would be cool if the launcher had save-game editing features and the like
   // Scan each save file, we inspect the `game-save`'s tag list.
@@ -189,19 +195,22 @@ pub async fn get_furthest_game_milestone(game_name: String) -> Result<String, Co
   // (need-resolution 7)
   // - byte 11 corresponds with it's task id
   // there is also a task status field but we don't really care about it, the task-status entry is sufficient
-  let game_save_dir = if let Some(config_dir) = config_dir() {
-    let expected_dir = config_dir.join("OpenGOAL").join(game_name).join("saves");
-    if !expected_dir.exists() {
-      info!(
-        "Expected save directory {} does not exist",
-        expected_dir.display()
-      );
+  let game_save_dir = match app_handle.path().config_dir() {
+    Ok(config_dir) => {
+      let expected_dir = config_dir.join("OpenGOAL").join(game_name).join("saves");
+      if !expected_dir.exists() {
+        info!(
+          "Expected save directory {} does not exist",
+          expected_dir.display()
+        );
+        return Ok("geyser".to_owned());
+      }
+      expected_dir
+    }
+    Err(_) => {
+      info!("Couldn't determine game save directory");
       return Ok("geyser".to_owned());
     }
-    expected_dir
-  } else {
-    info!("Couldn't determine game save directory");
-    return Ok("geyser".to_owned());
   };
 
   let milestones = get_jak1_milestones();
@@ -218,19 +227,16 @@ pub async fn get_furthest_game_milestone(game_name: String) -> Result<String, Co
     if let Some(ext) = entry.path().extension() {
       if ext == "bin" {
         info!("Scanning save {}", entry.path().display());
-        match get_saves_highest_milestone(&entry.into_path(), &milestones) {
-          Some((name, idx)) => {
-            info!("Furthest milestone {} at index {}", name, idx);
-            if idx > highest_milestone_idx {
-              highest_milestone_idx = idx;
-              furthest_milestone_name = name.to_owned();
-            }
+        if let Some((name, idx)) = get_saves_highest_milestone(&entry.into_path(), &milestones) {
+          info!("Furthest milestone {} at index {}", name, idx);
+          if idx > highest_milestone_idx {
+            highest_milestone_idx = idx;
+            furthest_milestone_name = name.to_owned();
           }
-          None => {}
         }
       }
     }
   }
 
-  return Ok(furthest_milestone_name);
+  Ok(furthest_milestone_name)
 }
