@@ -24,7 +24,7 @@
   import IconArrowDown from "~icons/mdi/arrow-down";
   import IconDelete from "~icons/mdi/delete";
   import IconInfo from "~icons/mdi/information";
-  import { convertFileSrc } from "@tauri-apps/api/tauri";
+  import { convertFileSrc } from "@tauri-apps/api/core";
   import {
     Accordion,
     AccordionItem,
