@@ -23,7 +23,7 @@
   import { generateSupportPackage } from "$lib/rpc/support";
   import { _ } from "svelte-i18n";
   import { emit } from "@tauri-apps/api/event";
-  import { arch, type } from "@tauri-apps/api/os";
+  import { arch, type } from "@tauri-apps/plugin-os";
   import { isMacOSVersion15OrAbove } from "$lib/rpc/util";
 
   export let activeGame: SupportedGame;
@@ -53,7 +53,7 @@
       // arm, we don't bother checking for simd
       // - if macOS (the only supported ARM platform), we check they are on atleast macOS 15
       // there is no easy way to check to see if they have rosetta 2, if you know of one, contribute it
-      if (osType !== "Darwin") {
+      if (osType !== "macos") {
         requirementsMet = false;
         return;
       }
@@ -61,7 +61,7 @@
       requirementsMet = macOSVersionSufficient && isOpenGLMet && isDiskSpaceMet;
     } else {
       const isAvxMet = await isAVXRequirementMet();
-      if (osType == "Windows_NT") {
+      if (osType == "windows") {
         const isVCCInstalled = await isMinimumVCCRuntimeInstalled();
         requirementsMet =
           isAvxMet && isOpenGLMet && isDiskSpaceMet && isVCCInstalled;
