@@ -1,5 +1,6 @@
 import { initLocales } from "$lib/i18n/i18n";
 import "./app.postcss";
+import { mount } from "svelte";
 import App from "./App.svelte";
 import { polyfillCountryFlagEmojis } from "country-flag-emoji-polyfill";
 
@@ -8,8 +9,6 @@ polyfillCountryFlagEmojis();
 // Register Translations
 initLocales(true);
 
-const app = new App({
-  target: document.getElementById("app"),
-});
+const app = mount(App, { target: document.getElementById("app") });
 
 export default app;
