@@ -1,4 +1,5 @@
 import { initLocales } from "$lib/i18n/i18n";
+import { mount } from "svelte";
 import App from "./Splash.svelte";
 import { polyfillCountryFlagEmojis } from "country-flag-emoji-polyfill";
 import "./splash.postcss";
@@ -7,7 +8,7 @@ import "./splash.postcss";
 export default (async () => {
   polyfillCountryFlagEmojis();
   await initLocales(false);
-  return new App({
+  return mount(App, {
     target: document.getElementById("app"),
   });
 })();
